refactor: Centralize the execution flow of World and fix bugs in Scheduler.
Centralize the execution of World to take charge of WorldRuntime, which removes the execution method of IWorldImpl.

Fix a bug in the Scheduler that did not apply a fixed ticktime.

+++ WITH_Server/WITH_Server_Framework/WorldRegistry.h
@@ -22,7 +22,7 @@ public:
 private:
 	struct WorldSlot
 	{
-		uint16 gen{ 0 };
+		uint32 gen{ 0 };
 		std::unique_ptr<IWorld> world{ nullptr };
 	};
